refactor: use a common Mul impl for all the unsigned integer gadgets
