+++ gadgets/src/traits/utilities/uint/arithmetic/mod.rs
@@ -20,6 +20,9 @@ pub use self::add::*;
 pub mod div;
 pub use self::div::*;
 
+pub mod mul;
+pub use self::mul::*;
+
 pub mod pow;
 pub use self::pow::*;
